requirements.txt
Django>=5.0.3
Pillow>=10.3.0
django-phonenumber-field>=7.3.0
django-colorfield>=0.11.0
phonenumbers>=8.13.36
